Consolidate catalogue item editor into single Swift file and JS file
Merge Page+CatalogueNew and Page+CatalogueItemEditor into a single
Page+CatalogueEditor.swift following the same pattern as album/song/book
editors. Move cataloguePreview page to Page+CatalogueItem.swift where
it belongs alongside CatalogueItemViewModel.

Merge catalogue-new.js and catalogue-item-edit.js into a single
catalogue-item-editor.js; create-only features (draft persistence, URL
autofill, preview button) are guarded behind an isNew check on the form
action so the same script drives both create and edit.
