+++ tmbr-web/Sources/App/Modules/Previews/Routes/Web/Pages/Page+CatalogueItem.swift
@@ -69,9 +69,46 @@ struct CatalogueItemViewModel: Encodable, Sendable {
 
 extension Template where Model == CatalogueItemViewModel {
     static let catalogueItem = Template(name: "Previews/preview")
+    static let cataloguePreview = Template(name: "Previews/preview")
+}
+
+private struct CataloguePreviewPayload: Content {
+    let title: String
+    let subtitle: String?
+    let artworkURL: String?
+    let url: String?
+    let notes: String
 }
 
 extension Page {
+    static var cataloguePreview: Self {
+        Page(template: .cataloguePreview) { request in
+            try await request.permissions.previews.create.grant()
+            let payload = try request.content.decode(CataloguePreviewPayload.self)
+            let formatter = MarkdownFormatter.html
+            let notes: [NoteViewModel] = payload.notes.isEmpty ? [] : [
+                NoteViewModel(
+                    id: UUID(),
+                    body: formatter.format(payload.notes),
+                    created: Date.now.formatted(.publishDate)
+                )
+            ]
+            let resource: Hyperlink? = payload.url.flatMap { urlString in
+                guard !urlString.isEmpty, let url = URL(string: urlString) else { return nil }
+                return Hyperlink(label: url.host ?? urlString, url: url)
+            }
+            return CatalogueItemViewModel(
+                title: "Preview: \(payload.title)",
+                subtitle: payload.subtitle,
+                artwork: payload.artworkURL.flatMap { url in
+                    url.isEmpty ? nil : ImageViewModel(previewURL: url)
+                },
+                notes: notes,
+                resources: resource.map { [$0] } ?? []
+            )
+        }
+    }
+
     static var catalogueItem: Self {
         Page(template: .catalogueItem) { request in
             guard let previewID = request.parameters.get("previewID", as: UUID.self) else {
